requirements.txt
streamlit
transformers
torch
pandas
textblob
streamlit-authenticator
pyyaml